Modificación distance_sensor para admitir uint16

+++ src/examples/px4_simple_app/px4_simple_app.cpp
@@ -177,7 +177,7 @@ void SensorDistancia::task_main()
 			(double)distance_sensor_param.current_distance,
 			(double)distance_sensor_param.covariance,
 			(uint8_t)distance_sensor_param.type,
-			(uint8_t)distance_sensor_param.orientation);
+			(uint16_t)distance_sensor_param.orientation);
 		}
 
 	}
